requirements.txt
Django==2.0.12
pandas==0.23.3
numpy==1.14.5
scikit-surprise==1.1.3